agent_system/requirements.txt
openai>=1.3.0
langchain>=0.0.300
langchain-openai>=0.0.1
autogen>=0.2.0
crewai>=0.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
redis>=5.0.0
pytest>=7.4.0
rich>=13.5.0 